requirements.txt
# Spyder OSINT 2_GUI - Dependencies
# Python 3.10+

# GUI (tkinter is built-in)
# Core dependencies
requests>=2.28.0
cryptography

# OSINT Tools
sherlock-project>=0.14.0
theHarvester>=4.0.0

# Data processing
tabulate>=0.9.0
beautifulsoup4>=4.11.0
selenium>=4.10.0

# Optional ML features:
# torch>=2.0.0
